feat: Add dimension columns to the figure outputs
