+++ packages/climate-ref/src/climate_ref/models/metric_value.py
@@ -2,12 +2,11 @@ import enum
 from collections.abc import Mapping
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from loguru import logger
-from sqlalchemy import Column, ForeignKey, Text, event
+from sqlalchemy import ForeignKey, event
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from climate_ref.models.base import Base, CreatedUpdatedMixin
-from climate_ref_core.pycmec.controlled_vocabulary import CV, Dimension
+from climate_ref.models.base import Base
+from climate_ref.models.mixins import CreatedUpdatedMixin, DimensionMixin
 
 if TYPE_CHECKING:
     from climate_ref.models.execution import Execution
@@ -27,7 +26,7 @@ class MetricValueType(enum.Enum):
     SERIES = "series"
 
 
-class MetricValue(CreatedUpdatedMixin, Base):
+class MetricValue(DimensionMixin, CreatedUpdatedMixin, Base):
     """
     Represents a single metric value
 
@@ -63,110 +62,8 @@ class MetricValue(CreatedUpdatedMixin, Base):
     This value is used to determine how the metric value should be interpreted.
     """
 
-    _cv_dimensions: ClassVar[list[str]] = []
-
-    @property
-    def dimensions(self) -> dict[str, str]:
-        """
-        Get the non-null dimensions and their values
-
-        Any changes to the resulting dictionary are not reflected in the object
-
-        Returns
-        -------
-            Collection of dimensions names and their values
-        """
-        dims = {}
-        for key in self._cv_dimensions:
-            value = getattr(self, key)
-            if value is not None:
-                dims[key] = value
-        return dims
-
     def __repr__(self) -> str:
         return f"<MetricValue id={self.id} execution={self.execution} dimensions={self.dimensions}>"
-
-    @staticmethod
-    def build_dimension_column(dimension: Dimension) -> Column[str]:
-        """
-        Create a column representing a CV dimension
-
-        These columns are not automatically generated with alembic revisions.
-        Any changes to this functionality likely require a manual database migration
-        of the existing columns.
-
-        Parameters
-        ----------
-        dimension
-            Dimension definition to create the column for.
-
-            Currently only the "name" field is being used.
-
-        Returns
-        -------
-            An instance of a sqlalchemy Column
-
-            This doesn't create the column in the database,
-            but enables the ORM to access it.
-
-        """
-        return Column(
-            dimension.name,
-            Text,
-            index=True,
-            nullable=True,
-            info={"skip_autogenerate": True},
-        )
-
-    @classmethod
-    def register_cv_dimensions(cls, cv: CV) -> None:
-        """
-        Register the dimensions supplied in the controlled vocabulary
-
-        This has to be done at run-time to support custom CVs.
-        Any extra columns already in the database, but not in the CV are ignored.
-
-        Parameters
-        ----------
-        cv
-            Controlled vocabulary being used by the application.
-            This controlled vocabulary contains the definitions of the dimensions that can be used.
-        """
-        for dimension in cv.dimensions:
-            target_attribute = dimension.name
-            if target_attribute in cls._cv_dimensions:
-                continue
-
-            cls._cv_dimensions.append(target_attribute)
-            logger.debug(f"Registered MetricValue dimension: {target_attribute}")
-
-            if hasattr(cls, target_attribute):
-                # This should only occur in test suite as we don't support removing dimensions at runtime
-                logger.warning("Column attribute already exists on MetricValue. Ignoring")
-            else:
-                setattr(cls, target_attribute, cls.build_dimension_column(dimension))
-
-            # TODO: Check if the underlying table already contains columns
-
-    @classmethod
-    def _reset_cv_dimensions(cls) -> None:
-        """
-        Remove any previously registered dimensions
-
-        Used by the test suite and should not be called at runtime.
-
-        This doesn't remove any previous column definitions due to a limitation that columns in
-        declarative classes cannot be removed.
-        This means that `hasattr(MetricValue, "old_attribute")`
-        will still return True after resetting, but the values will not be included in any executions.
-        """
-        logger.warning(f"Removing MetricValue dimensions: {cls._cv_dimensions}")
-
-        keys = list(cls._cv_dimensions)
-        for key in keys:
-            cls._cv_dimensions.remove(key)
-
-        assert not len(cls._cv_dimensions)
 
 
 class ScalarMetricValue(MetricValue):
